+++ pkg/gcp/client/compute.go
@@ -10,7 +10,6 @@ import (
 	"net/http"
 	"strings"
 
-	"golang.org/x/oauth2"
 	"google.golang.org/api/compute/v1"
 	"google.golang.org/api/googleapi"
 	"google.golang.org/api/option"
@@ -101,12 +100,11 @@ type computeClient struct {
 // Delete operations will ignore errors when the respective resource can not be found, meaning that the Delete operations will never return HTTP 404 errors.
 // Update operations will ignore errors when the update operation is a no-op, meaning that Update operations will ignore HTTP 304 errors.
 func NewComputeClient(ctx context.Context, credentialsConfig *gcp.CredentialsConfig) (ComputeClient, error) {
-	tokenSource, err := tokenSource(ctx, credentialsConfig, []string{compute.ComputeScope})
+	httpClient, err := httpClient(ctx, credentialsConfig, []string{compute.ComputeScope})
 	if err != nil {
 		return nil, err
 	}
 
-	httpClient := oauth2.NewClient(ctx, tokenSource)
 	service, err := compute.NewService(ctx, option.WithHTTPClient(httpClient))
 	if err != nil {
 		return nil, err
